+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 23-pre
 
+- Added:
+	- query tools-usage (like tools-usage-per-month but for a year or for whole history) by @lldelisle
+- Fixed:
+	- query tools-usage-per-month when 'no_version' was used there were still one line per version.
+
 # 22
 
 - Added: